requirements.txt
discord.py==2.3.2
pytest==7.4.3
python-dotenv==1.0.0 